Move s3 utils into pipeline service
